move functions only needed for testing out of the Server class
