+++ src/test/java/de/komoot/photon/elasticsearch/ImporterTest.java
@@ -1,10 +1,9 @@
 package de.komoot.photon.elasticsearch;
 
 import de.komoot.photon.ESBaseTester;
-import de.komoot.photon.PhotonDoc;
 import de.komoot.photon.Importer;
+import de.komoot.photon.PhotonDoc;
 import de.komoot.photon.searcher.PhotonResult;
-import org.elasticsearch.action.get.GetResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
