+++ src/test/java/de/komoot/photon/elasticsearch/UpdaterTest.java
@@ -1,11 +1,10 @@
 package de.komoot.photon.elasticsearch;
 
 import de.komoot.photon.ESBaseTester;
-import de.komoot.photon.PhotonDoc;
 import de.komoot.photon.Importer;
+import de.komoot.photon.PhotonDoc;
 import de.komoot.photon.Updater;
 import de.komoot.photon.searcher.PhotonResult;
-import org.elasticsearch.action.get.GetResponse;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
